refactor(pgmq-rs): inline PGMQueueExt struct into pg_ext/mod.rs
The `pg_ext::deprecated` submodule was introduced when PGMQueueExt was
moved out of `pg_ext/mod.rs`; the module-name "deprecated" was a value
judgment in the path. Restore the pre-b1cb6ad layout: the struct lives
directly in `pg_ext/mod.rs` again, no submodule.

Public paths unchanged — `pgmq::PGMQueueExt` and `pgmq::pg_ext::PGMQueueExt`
still resolve. The struct itself keeps its `#[deprecated]` attribute; only
the file/module location changes.
